+++ packages/react/src/withTheme.tsx
@@ -6,23 +6,22 @@ import { WithThemeWrappedProps, WithThemeWrapperProps, WrapperComponent } from '
 /**
  * Wrap a React component with an HOC that injects the current theme object as a prop.
  */
-export default function withTheme<Props extends object = {}>(
-  WrappedComponent: React.ComponentType<Props & WithThemeWrappedProps>,
-): React.FunctionComponent<Props & WithThemeWrapperProps> &
-  WrapperComponent<Props & WithThemeWrappedProps> {
-  const baseName = WrappedComponent.displayName || WrappedComponent.name;
+export default function withTheme() /* infer */ {
+  return function withThemeComposer<Props extends object = {}>(
+    WrappedComponent: React.ComponentType<Props & WithThemeWrappedProps>,
+  ): React.FunctionComponent<Props & WithThemeWrapperProps> & WrapperComponent {
+    function WithTheme({ wrappedRef, ...props }: Props & WithThemeWrapperProps) {
+      const theme = useTheme();
 
-  function WithTheme({ wrappedRef, ...props }: Props & WithThemeWrapperProps) {
-    const theme = useTheme();
+      return <WrappedComponent {...(props as any)} ref={wrappedRef} theme={theme} />;
+    }
 
-    return <WrappedComponent {...(props as any)} ref={wrappedRef} theme={theme} />;
-  }
+    hoistNonReactStatics(WithTheme, WrappedComponent);
 
-  hoistNonReactStatics(WithTheme, WrappedComponent);
+    WithTheme.displayName = `withTheme(${WrappedComponent.displayName || WrappedComponent.name})`;
 
-  WithTheme.displayName = `withTheme(${baseName})`;
+    WithTheme.WrappedComponent = WrappedComponent;
 
-  WithTheme.WrappedComponent = WrappedComponent;
-
-  return WithTheme;
+    return WithTheme;
+  };
 }