+++ packages/style/tests/helpers/applyUnitToValue.test.ts
@@ -2,25 +2,29 @@ import applyUnitToValue from '../../src/helpers/applyUnitToValue';
 
 describe('applyUnitToValue()', () => {
   it('does nothing to strings', () => {
-    expect(applyUnitToValue('display', 'foo')).toBe('foo');
-    expect(applyUnitToValue('display', '100px')).toBe('100px');
-    expect(applyUnitToValue('display', 'red')).toBe('red');
+    expect(applyUnitToValue('display', 'foo', 'px')).toBe('foo');
+    expect(applyUnitToValue('display', '100px', 'px')).toBe('100px');
+    expect(applyUnitToValue('display', 'red', 'px')).toBe('red');
   });
 
   it('does nothing to zeros', () => {
-    expect(applyUnitToValue('width', 0)).toBe('0');
+    expect(applyUnitToValue('width', 0, 'px')).toBe('0');
   });
 
   it('does nothing to unitless properties', () => {
-    expect(applyUnitToValue('lineHeight', 1.5)).toBe('1.5');
+    expect(applyUnitToValue('lineHeight', 1.5, 'px')).toBe('1.5');
   });
 
   it('does nothing if a unit already exists', () => {
-    expect(applyUnitToValue('width', '100em')).toBe('100em');
-    expect(applyUnitToValue('height', '10vh')).toBe('10vh');
+    expect(applyUnitToValue('width', '100em', 'px')).toBe('100em');
+    expect(applyUnitToValue('height', '10vh', 'px')).toBe('10vh');
   });
 
   it('appends px to numbers', () => {
-    expect(applyUnitToValue('width', 100)).toBe('100px');
+    expect(applyUnitToValue('width', 100, 'px')).toBe('100px');
+  });
+
+  it('can customize unit', () => {
+    expect(applyUnitToValue('width', 100, 'em')).toBe('100em');
   });
 });