+++ packages/utils/src/hyphenate.ts
@@ -1,4 +1,5 @@
-const PATTERN = /[A-Z]/gu;
+const CAMEL_CASE_PATTERN = /[A-Z]/gu;
+const VENDOR_PREFIX_PATTERN = /^(ms|moz|webkit)/u;
 const cache: { [key: string]: string } = {};
 
 function toLower(match: string): string {
@@ -7,7 +8,13 @@ function toLower(match: string): string {
 
 export default function hyphenate(value: string): string {
   if (!cache[value]) {
-    cache[value] = value.replace(PATTERN, toLower);
+    let result = value.replace(CAMEL_CASE_PATTERN, toLower);
+
+    if (VENDOR_PREFIX_PATTERN.test(result)) {
+      result = `-${result}`;
+    }
+
+    cache[value] = result;
   }
 
   return cache[value];