+++ packages/utils/tests/hyphenate.test.ts
@@ -8,4 +8,10 @@ describe('hyphenate()', () => {
   it('doesnt convert already hyphenated strings', () => {
     expect(hyphenate('foo-bar-baz')).toBe('foo-bar-baz');
   });
+
+  it('supports vendor prefixed properties', () => {
+    expect(hyphenate('WebkitTransition')).toBe('-webkit-transition');
+    expect(hyphenate('MozBoxSizing')).toBe('-moz-box-sizing');
+    expect(hyphenate('MsFontSmoothing')).toBe('-ms-font-smoothing');
+  });
 });